group clone controller tests moved
